+++ wled00/cfg.cpp
@@ -279,11 +279,13 @@ void deserializeConfig() {
     CJSON(timerMacro[it], timer[F("macro")]);
 
     byte dowPrev =  timerWeekday[it];
-    bool actPrev = timerWeekday[it] & 0x01;
+    //note: act is currently only 0 or 1.
+    //the reason we are not using bool is that the on-disk type in 0.11.0 was already int
+    int actPrev = timerWeekday[it] & 0x01;
     CJSON(timerWeekday[it], timer[F("dow")]);
     if (timerWeekday[it] != dowPrev) { //present in JSON
       timerWeekday[it] <<= 1; //add active bit
-      bool act = timer[F("en")] | actPrev;
+      int act = timer[F("en")] | actPrev;
       if (act) timerWeekday[it]++;
     }