+++ wled00/cfg.cpp
@@ -96,6 +96,7 @@ void deserializeConfig() {
   CJSON(strip.ablMilliampsMax, hw_led[F("maxpwr")]);
   CJSON(strip.milliampsPerLed, hw_led[F("ledma")]);
   CJSON(strip.reverseMode, hw_led[F("rev")]);
+  CJSON(strip.rgbwMode, hw_led[F("rgbwm")]);
 
   JsonObject hw_led_ins_0 = hw_led[F("ins")][0];
   //bool hw_led_ins_0_en = hw_led_ins_0[F("en")]; // true
@@ -172,7 +173,7 @@ void deserializeConfig() {
   CJSON(presetCycleMax, def_cy[F("range")][1]);
 
   tdd = def_cy[F("dur")] | -1;
-  if (tdd >= 0) presetCycleTime = tdd * 100;
+  if (tdd > 0) presetCycleTime = tdd;
 
   // if   == interfaces
   JsonObject interfaces = doc["if"];
@@ -263,7 +264,11 @@ void deserializeConfig() {
   CJSON(countdownMode, ol[F("cntdwn")]);
   overlayCurrent = overlayDefault;
 
-  JsonArray ol_cntdwn = ol[F("cntdwn")]; //[20,12,31,23,59,59]
+  CJSON(overlayMin, ol[F("min")]);
+  CJSON(overlayMax, ol[F("max")]);
+  CJSON(analogClock12pixel, ol[F("o12pix")]);
+  CJSON(analogClock5MinuteMarks, ol[F("o5m")]);
+  CJSON(analogClockSecondsTrail, ol[F("osec")]);
 
   //timed macro rules
   JsonObject tm = doc[F("timers")];
@@ -286,11 +291,13 @@ void deserializeConfig() {
     CJSON(timerMacro[it], timer[F("macro")]);
 
     byte dowPrev =  timerWeekday[it];
-    bool actPrev = timerWeekday[it] & 0x01;
+    //note: act is currently only 0 or 1.
+    //the reason we are not using bool is that the on-disk type in 0.11.0 was already int
+    int actPrev = timerWeekday[it] & 0x01;
     CJSON(timerWeekday[it], timer[F("dow")]);
     if (timerWeekday[it] != dowPrev) { //present in JSON
       timerWeekday[it] <<= 1; //add active bit
-      bool act = timer[F("en")] | actPrev;
+      int act = timer[F("en")] | actPrev;
       if (act) timerWeekday[it]++;
     }
 
@@ -406,6 +413,7 @@ void serializeConfig() {
   hw_led[F("maxpwr")] = strip.ablMilliampsMax;
   hw_led[F("ledma")] = strip.milliampsPerLed;
   hw_led[F("rev")] = strip.reverseMode;
+  hw_led[F("rgbwm")] = strip.rgbwMode;
 
   JsonArray hw_led_ins = hw_led.createNestedArray("ins");
 
@@ -511,7 +519,7 @@ void serializeConfig() {
     JsonArray def_cy_range = def_cy.createNestedArray("range");
     def_cy_range.add(presetCycleMin);
     def_cy_range.add(presetCycleMax);
-    def_cy[F("dur")] = presetCycleTime / 100;
+    def_cy[F("dur")] = presetCycleTime;
   }
 
   JsonObject interfaces = doc.createNestedObject("if");
@@ -594,6 +602,12 @@ void serializeConfig() {
   JsonObject ol = doc.createNestedObject("ol");
   ol[F("clock")] = overlayDefault;
   ol[F("cntdwn")] = countdownMode;
+
+  ol[F("min")] = overlayMin;
+  ol[F("max")] = overlayMax;
+  ol[F("o12pix")] = analogClock12pixel;
+  ol[F("o5m")] = analogClock5MinuteMarks;
+  ol[F("osec")] = analogClockSecondsTrail;
 
   JsonObject timers = doc.createNestedObject("timers");