+++ wled00/led.cpp
@@ -94,35 +94,47 @@ void colorUpdated(int callMode)
 {
   //call for notifier -> 0: init 1: direct change 2: button 3: notification 4: nightlight 5: other (No notification)
   //                     6: fx changed 7: hue 8: preset cycle 9: blynk 10: alexa
-  if (callMode != NOTIFIER_CALL_MODE_INIT && 
-      callMode != NOTIFIER_CALL_MODE_DIRECT_CHANGE && 
+  if (callMode != NOTIFIER_CALL_MODE_INIT &&
+      callMode != NOTIFIER_CALL_MODE_DIRECT_CHANGE &&
       callMode != NOTIFIER_CALL_MODE_NO_NOTIFY) strip.applyToAllSelected = true; //if not from JSON api, which directly sets segments
-  
+
+  bool someSel = false;
+
+  if (callMode == NOTIFIER_CALL_MODE_NOTIFICATION) {
+    someSel = (receiveNotificationBrightness || receiveNotificationColor || receiveNotificationEffects);
+  }
+
+  //Notifier: apply received FX to selected segments only if actually receiving FX
+  if (someSel) strip.applyToAllSelected = receiveNotificationEffects;
+
   bool fxChanged = strip.setEffectConfig(effectCurrent, effectSpeed, effectIntensity, effectFFT1, effectFFT2, effectFFT3, effectPalette);
   bool colChanged = colorChanged();
+
+  //Notifier: apply received color to selected segments only if actually receiving color
+  if (someSel) strip.applyToAllSelected = receiveNotificationColor;
 
   if (fxChanged || colChanged)
   {
     if (realtimeTimeout == UINT32_MAX) realtimeTimeout = 0;
     if (isPreset) {isPreset = false;}
         else {currentPreset = -1;}
-        
+
     notify(callMode);
-    
+
     //set flag to update blynk and mqtt
-    if (callMode != NOTIFIER_CALL_MODE_PRESET_CYCLE) interfaceUpdateCallMode = callMode;
+    interfaceUpdateCallMode = callMode;
   } else {
-    if (nightlightActive && !nightlightActiveOld && 
-        callMode != NOTIFIER_CALL_MODE_NOTIFICATION && 
+    if (nightlightActive && !nightlightActiveOld &&
+        callMode != NOTIFIER_CALL_MODE_NOTIFICATION &&
         callMode != NOTIFIER_CALL_MODE_NO_NOTIFY)
     {
-      notify(NOTIFIER_CALL_MODE_NIGHTLIGHT); 
+      notify(NOTIFIER_CALL_MODE_NIGHTLIGHT);
       interfaceUpdateCallMode = NOTIFIER_CALL_MODE_NIGHTLIGHT;
     }
   }
-  
+
   if (!colChanged) return; //following code is for e.g. initiating transitions
-  
+
   if (callMode != NOTIFIER_CALL_MODE_NO_NOTIFY && nightlightActive && (nightlightMode == NL_MODE_FADE || nightlightMode == NL_MODE_COLORFADE))
   {
     briNlT = bri;
@@ -145,14 +157,14 @@ void colorUpdated(int callMode)
 
   //deactivate nightlight if target brightness is reached
   if (bri == nightlightTargetBri && callMode != NOTIFIER_CALL_MODE_NO_NOTIFY && nightlightMode != NL_MODE_SUN) nightlightActive = false;
-  
+
   if (fadeTransition)
   {
     //set correct delay if not using notification delay
     if (callMode != NOTIFIER_CALL_MODE_NOTIFICATION && !jsonTransitionOnce) transitionDelayTemp = transitionDelay;
     jsonTransitionOnce = false;
     if (transitionDelayTemp == 0) {setLedsStandard(); strip.trigger(); return;}
-    
+
     if (transitionActive)
     {
       for (byte i=0; i<4; i++)
@@ -183,7 +195,7 @@ void updateInterfaces(uint8_t callMode)
     espalexaDevice->setColor(col[0], col[1], col[2]);
   }
   #endif
-  if (callMode != NOTIFIER_CALL_MODE_BLYNK && 
+  if (callMode != NOTIFIER_CALL_MODE_BLYNK &&
       callMode != NOTIFIER_CALL_MODE_NO_NOTIFY) updateBlynk();
   doPublishMqtt = true;
   lastInterfaceUpdate = millis();
@@ -199,7 +211,7 @@ void handleTransitions()
     interfaceUpdateCallMode = 0; //disable
   }
   if (doPublishMqtt) publishMqtt();
-  
+
   if (transitionActive && transitionDelayTemp > 0)
   {
     float tper = (millis() - transitionStartTime)/(float)transitionDelayTemp;
@@ -219,7 +231,7 @@ void handleTransitions()
       colSecT[i] = colSecOld[i]+((colSec[i] - colSecOld[i])*tper);
     }
     briT    = briOld   +((bri    - briOld   )*tper);
-    
+
     setAllLeds();
   }
 }
@@ -306,10 +318,10 @@ void handleNightlight()
     if (bri == 0 || nightlightActive) return;
 
     if (presetCycCurr < presetCycleMin || presetCycCurr > presetCycleMax) presetCycCurr = presetCycleMin;
-    applyPreset(presetCycCurr);
+    applyPreset(presetCycCurr); //this handles colorUpdated() for us
     presetCycCurr++;
     if (presetCycCurr > 250) presetCycCurr = 1;
-    colorUpdated(NOTIFIER_CALL_MODE_PRESET_CYCLE);
+    interfaceUpdateCallMode = 0; //disable updates to MQTT and Blynk
   }
 }