+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ### Development versions after 0.11.0 release
 
+#### Build 2012150
+
+-   Added Blends FX (PR #1491)
+-   Fixed an issue that made it impossible to deactivate timed presets
+
 #### Build 2012140
 
 -   Added Preset ID quick display option (PR #1462)